+++ InventoryToolsMock/Program.cs
@@ -2,6 +2,8 @@
 using DalaMock.Core.DI;
 using DalaMock.Core.Mocks;
 using Dalamud.Bindings.ImGui;
+using Dalamud.Game;
+using Lumina.Data;
 using Lumina.Excel.Sheets;
 
 namespace InventoryToolsMock
@@ -10,7 +12,9 @@ namespace InventoryToolsMock
     {
         static void Main(string[] args)
         {
-            var mockContainer = new MockContainer();
+            var mockContainer = new MockContainer(new MockDalamudConfiguration()
+            {
+            });
             var mockDalamudUi = mockContainer.GetMockUi();
             var pluginLoader = mockContainer.GetPluginLoader();
             mockContainer.GetGameData().GetExcelSheet<Item>();